+++ nata-monkey/src/main/java/com/nata/monkeys/AbstractMonkey.java
@@ -1,16 +1,21 @@
 package com.nata.monkeys;
 
+import com.nata.action.Action;
+import com.nata.action.TextInputAction;
 import com.nata.cmd.AdbDevice;
 import com.nata.element.DumpService;
 import com.nata.element.Element;
 import com.nata.element.UINode;
+import com.nata.rules.Rule;
+import com.nata.rules.RuleParser;
+import com.nata.rules.Rules;
 import com.nata.state.State;
 import org.dom4j.DocumentException;
+import org.dom4j.Text;
 import org.xml.sax.SAXException;
 
 import java.io.*;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 
 /**
  * Author: Calvin Meng
@@ -22,6 +27,8 @@ public abstract class AbstractMonkey {
     private final String pkg;
     private final String act;
     private final AdbDevice device;
+    private Set<UINode> widgetSet = new HashSet<>();
+    private Map<String,String> knowledge = new HashMap<>();
 
     public AbstractMonkey(String name, String pkg, String act, AdbDevice device) {
         this.name = name;
@@ -35,35 +42,33 @@ public abstract class AbstractMonkey {
         return pkgAct;
     }
 
-    public void clearAppData(){
-       device.clearAppData(pkg);
+    public void clearAppData() {
+        device.clearAppData(pkg);
     }
-
-
 
 
     public State getCurrentState() {
         String curActivity = getCurrentActivity();
-        String appPackage  = getCurrentPackage();
-        State state = new State(appPackage,curActivity);
+        String appPackage = getCurrentPackage();
+        State state = new State(appPackage, curActivity);
         List<UINode> uiNodes = GrabCurrentUi();
 
 
         // Add the uinodes with resourceId;
-        for(UINode node : uiNodes){
-            if(!node.getResourceId().equals("")){
-               state.addUINode(node);
+        for (UINode node : uiNodes) {
+            if (!node.getResourceId().equals("")) {
+                state.addUINode(node);
             }
         }
 
         return state;
     }
 
-    public String getCurrentPackage(){
+    public String getCurrentPackage() {
         return device.getCurrentPackageName();
     }
 
-    public String getCurrentActivity(){
+    public String getCurrentActivity() {
         return device.getCurrentActivity();
     }
 
@@ -82,9 +87,10 @@ public abstract class AbstractMonkey {
         File dumpFile = device.dumpUI();
         try {
             List<UINode> list = DumpService.getNodes(dumpFile);
-//            for (UINode node : list) {
+            for (UINode node : list) {
 //                System.out.println(node);
-//            }
+                widgetSet.add(node);
+            }
             return list;
         } catch (DocumentException e) {
             e.printStackTrace();
@@ -99,14 +105,14 @@ public abstract class AbstractMonkey {
         return null;
     }
 
+    protected Set<UINode> getWidgetSet(){
+        return widgetSet;
+    }
+
     //TODO: detect errors,how?
     public boolean isCrashed() {
         return false;
     }
-
-//    public void SleepShort() {
-//        device.sleep(500);
-//    }
 
     public String getName() {
         return name;
@@ -120,6 +126,33 @@ public abstract class AbstractMonkey {
         return pkg;
     }
 
+
+    public void brainStorm(Action action){
+        if(action instanceof TextInputAction){
+            TextInputAction textInputAction = (TextInputAction) action;
+            String resourceId = textInputAction.getElement().getResourceId();
+            String value = knowledge.get(resourceId);
+            if(value != null){
+                textInputAction.setText(value);
+            }
+        }
+    }
+    public void learn(String rulePath){
+        File file = new File(rulePath);
+        Rules rules = RuleParser.parse(file);
+        learn(rules);
+    }
+
+    public void learn(Rules rules){
+        List<Rule> ruleList = rules.getRules();
+        for (Rule rule:ruleList) {
+            knowledge.put(rule.getResouceId(),rule.getValue());
+        }
+    }
+
+    public void learn(Rule rule){
+        knowledge.put(rule.getResouceId(),rule.getValue());
+    }
     /**
      * The monkeys know how to play with the device
      */
@@ -134,4 +167,5 @@ public abstract class AbstractMonkey {
      * They can report the running status after they stop
      */
     public abstract void report();
+
 }