+++ nata-monkey/src/main/java/com/nata/rules/Rule.java
@@ -1,0 +1,27 @@
+package com.nata.rules;
+
+/**
+ * Author: Calvin Meng
+ * Blog: mclspace.com  Email: rdmclin2@gamil.com
+ * Update: 2016-03-30 21:13
+ */
+public class Rule {
+    private String resouceId;
+    private String value;
+
+    public String getResouceId() {
+        return resouceId;
+    }
+
+    public void setResouceId(String resouceId) {
+        this.resouceId = resouceId;
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public void setValue(String value) {
+        this.value = value;
+    }
+}