+++ nata-monkey/src/main/java/com/nata/rules/Rules.java
@@ -1,0 +1,35 @@
+package com.nata.rules;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Author: Calvin Meng
+ * Blog: mclspace.com  Email: rdmclin2@gamil.com
+ * Update: 2016-03-30 21:01
+ */
+public class Rules {
+    private String type;
+    private List<Rule> rules  = new ArrayList<>();
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public List<Rule> getRules() {
+        return rules;
+    }
+
+    public void setRules(List<Rule> rules) {
+        this.rules = rules;
+    }
+
+    public void addRule(Rule rule){
+        rules.add(rule);
+    }
+
+}